requirements.txt
streamlit==1.28.0
PyPDF2==3.0.1
python-docx==1.0.1
groq==0.4.2
python-dotenv==1.0.0
thefuzz[speedup]==0.22.1
python-Levenshtein==0.21.1
numpy==1.24.0
